+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased ()
 -------------
 
+- Fixed aggWindow snapping in all cases, always.
+
 - Prettier zoom buttons in timeline.
 
 - Fixed onload error accessing layers before availability.